requirements.txt
fastapi
uvicorn
chromadb
openai
python-docx
requests
pydantic
